take some asset classes out of ViewerEnvironment.cs to own files
